Biologos terminados

+++ Models/BiologoDTO.cs
@@ -1,5 +1,3 @@
-using System.ComponentModel.DataAnnotations;
-
 namespace BettleHubCsharp.Models;
 
 public class BiologoDTO{
@@ -7,8 +5,7 @@ public class BiologoDTO{
     public required string Correo_biologo { get; set; }
     public required string Nombre_biologo { get; set; }
     public required int? Edad_biologo { get; set; }
-    public required int? Telefono_biologo { get; set; }
+    public required long? Telefono_biologo { get; set; }
     public required string Contrasena_biologo { get; set; }
     public required string Usuario_biologo { get; set; }
-    
 }
